[gemini]
Let's modify this so that, if the chunk appears to already contain an
OSC8 hyperlink, then we bail out and just return the chunk verbatim.
And let's apply that logic even if the chunk only contains the
beginning, or the end of an OSC8 hyperlink: anything that's a sure
sign it's part of an OSC8 hyperlink.

Okay, that's a good safeguard to prevent accidentally creating nested
or broken hyperlinks if the underlying program already generates them.
I'll modify the transform function in src/transform.rs. Before calling
transform_and_reapply_ansi, I'll add a check to see if the input chunk
contains the OSC 8 hyperlink introducer \x1b]8;;. If it does, the
function will return the original chunk immediately as a String.
Otherwise, it will proceed with the transformation as usual.