Fix exit hang: don't join input_thread
input_thread blocks on stdin.read() which won't EOF until the
controlling terminal closes. Waiting for it to join caused xolmis
to hang on exit when run non-interactively (e.g. in tmux scripts).

The existing comment acknowledged this but the join was still called
before process::exit. Now we just let process::exit terminate it.